pyproject.toml
[tool.poetry]
name = "wanguard-bgp-adapter"
version = "0.1.0"
description = ""
authors = ["pppwaw <pppwaw@pppwaw.top>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.63.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.63.0"
ruff = "^0.4.2"

[tool.ruff]
line-length = 120
# https://beta.ruff.rs/docs/rules/
# Exclude a variety of commonly ignored directories.
respect-gitignore = true
[tool.ruff.lint]
ignore-init-module-imports = true
select = ["E", "W", "F"]
ignore = ["F401"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
